combine integration test folders
